extract external dns list
