+++ src/components/status-badges/index.ts
@@ -1,6 +1,7 @@
 export * from './build-secret-status-badge';
 export * from './component-secret-status-badge';
 export * from './component-status-badge';
+export * from './external-dns-status-badge';
 export * from './generic-status-badge';
 export * from './image-hub-secret-status-badge';
 export * from './pipeline-run-status-badge';
